pyproject.toml
[project]
name = "rag-exper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "ipython>=8.31.0",
    "langchain-community>=0.3.16",
    "langchain-groq>=0.2.4",
    "langchain-nomic>=0.1.4",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.5",
    "langchain>=0.3.16",
    "llama-index>=0.12.14",
    "llmsherpa>=0.1.4",
    "locust>=2.32.8",
    "nomic[local]>=3.4.1",
    "pypdf>=5.2.0",
    "python-dotenv>=1.0.1",
]
